Front-end progress
Changes in how light/dark theme is implemented.
Creation of a page that will later be the profile page.
Implementation of a bottom navigation bar to change between current task list and task history.
Implementation of the routing for all current pages (register and login are redirecting without authentication for testing purpouses).